Correctly deal with a virtualenv python executable with a broken code signature.
This is an issue with virtualenv on macOS, because the Python binary copied into the virtualenv is code signed as of 10.12 when using Apple's python. Running that python binary outside the python framework bundle means the associated code signature (stored in the bundle) is not found, and because the binary still has the LC_CODE_SIGNATURE load command, code signature validation is attempted and fails. This causes platform security software to reject this python when it attempts to interact with keychain. For our keychain interactions, no specific code signature or entitlements are required - in fact, it's ok to have *no* code signature. What isn't ok is an *invalid* code signature. To work around this, we check whether the virtualenv python has a broken code signature, and if it does, we ad-hoc resign it.