Check to ensure PSQL is available running command (#480)
* Check to ensure PSQL is available running command

Tool currently crashes if PSQL isn't found in the path specified. Should gracefully exit with an error instead.

* Update dependencies and add note about bash

This list of dependencies is the full list of what's needed on a fresh Ubuntu 16.04.1 Docker image. Added a note about bash since the default shell that is used for a script from the command-line may be /bin/sh on some distributions, which is different than Debian)